카드와 덱 파일 생성
햇갈려서 분리했습니다.
