+++ CardGameApp/CardGameApp/Model/Card.swift
@@ -1,14 +1,13 @@
 //
-//  CardModel.swift
+//  Card.swift
 //  CardGameApp
 //
-//  Created by Cloud on 2020/02/10.
+//  Created by Cloud on 2020/02/28.
 //  Copyright © 2020 Cloud. All rights reserved.
 //
 
 import Foundation
 
-//MARK: - Card
 class Card {
     
     enum suit: CaseIterable {
@@ -31,56 +30,12 @@ class Card {
     }
 }
 
-//MARK: - Deck
-struct Deck {
-    
-    private var cards: [Card] = []
-    
-    init() {
-        makeDeck()
-    }
-    
-    mutating func shuffle() {
-        cards.shuffle()
-    }
-    
-    mutating func reset() {
-        makeDeck()
-    }
-    
-    mutating func removeOne() -> Card? {
-        cards.popLast()
-    }
-    
-    private func makeSuit(suit: Card.suit) -> [Card] {
-        return (1...13).map {
-            Card(suit: suit, rank: Card.rank(rawValue: $0)!)
-        }
-    }
-    
-    private mutating func makeDeck() {
-        Card.suit.allCases.forEach {
-            cards.append(contentsOf: makeSuit(suit: $0))
-        }
-        cards.shuffle()
-    }
-}
-
-//MARK: - Extension
-//MARK: Card
 extension Card: CustomStringConvertible {
     var description: String {
         return self.suit.description + self.rank.description
     }
 }
 
-extension Card: Equatable {
-    static func == (lhs: Card, rhs: Card) -> Bool {
-        return lhs.suit == rhs.suit && lhs.rank == rhs.rank
-    }
-}
-
-//MARK: Card.rank
 extension Card.rank: CustomStringConvertible {
     var description: String {
         switch self {
@@ -93,7 +48,6 @@ extension Card.rank: CustomStringConvertible {
     }
 }
 
-//MARK: Card.suit
 extension Card.suit: CustomStringConvertible {
     var description: String {
         switch self {
@@ -102,12 +56,5 @@ extension Card.suit: CustomStringConvertible {
         case .heart: return "h"
         case .spade: return "s"
         }
-    }
-}
-
-//MARK: - Deck
-extension Deck: Equatable {
-    static func == (lhs: Deck, rhs: Deck) -> Bool {
-        return lhs.cards == rhs.cards
     }
 }
